server/requirements.txt
fastapi
uvicorn
pymongo
python-multipart
passlib[bcrypt]
bcrypt
kafka-python
python-jose[cryptography]
jinja2